+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Properly rewrite redirect target url when present in <meta> HTML tag (#237)
+
 ### Changed
 
 - Generate fuzzy rules tests in Python and Javascript (#284)